# Improve performance of data onboarding