@gammazero 2025-04-02
Revised with improved solution:

Input to the queue is buffered in memory. The contents of the buffer are written to the datastore when the input buffer contains batchSize items, or when the queue has been idel (no input or output) for a long enough time. CIDs to dequeue are read, in order, from the input buffer when there are none in the datastore. Otherwise they are read from the datastore.

Reading queued items from the input can allow CIDs to remain in memory, since the queue is not idle and the input buffer size if reduced by reading items from it. When the queue is closed, any remaining items in memory are written to the datastore.

This gives us in-memory behavior until there are too many CIDs in memory, or until items have been sitting idle in memory too long.